Fix the metadata generated after a call to keys_to_xxx
