+++ include/equistore.h
@@ -542,6 +542,11 @@ eqs_status_t eqs_tensormap_components_to_properties(struct eqs_tensormap_t *tens
  * be merged together. The resulting merged blocks will have `variables` as
  * the last sample variables, preceded by the current samples.
  *
+ * The order of the samples is controlled by `sort_samples`. If
+ * `sort_samples` is true, samples are re-ordered to keep them
+ * lexicographically sorted. Otherwise they are kept in the order in which
+ * they appear in the blocks.
+ *
  * This function is only implemented if all merged block have the same
  * property labels.
  *
@@ -551,6 +556,8 @@ eqs_status_t eqs_tensormap_components_to_properties(struct eqs_tensormap_t *tens
  * @param tensor pointer to an existing tensor map
  * @param variables names of the key variables to move to the samples
  * @param variables_count number of entries in the `variables` array
+ * @param sort_samples whether to sort the samples lexicographically after
+ *                     merging blocks or not
  *
  * @returns The status code of this operation. If the status is not
  *          `EQS_SUCCESS`, you can use `eqs_last_error()` to get the full
@@ -558,7 +565,8 @@ eqs_status_t eqs_tensormap_components_to_properties(struct eqs_tensormap_t *tens
  */
 eqs_status_t eqs_tensormap_keys_to_samples(struct eqs_tensormap_t *tensor,
                                            const char *const *variables,
-                                           uint64_t variables_count);
+                                           uint64_t variables_count,
+                                           bool sort_samples);
 
 #ifdef __cplusplus
 } // extern "C"
